Add support for James Tucker's async response scheme. Added chunked responses.
Check out
http://github.com/raggi/thin/tree/async_for_rack/example/async_app.ru
for an example application using Thin and EventMachine
With Flow you can use EV watcher callbacks or you can yield the Fiber.
James, if you're reading, thank you for taking the time to explain this to
me. Forgive my sloth-like mind.

Additionally I added support for chunked responses which should be used if
you do not know the content length of your message. Set a header
"Content-Encoding" => "chunked"
(This is also subject to change)